Adjust webhooks to multi validation handlers
