+++ pkg/webhook/servicecatalog/servicebinding/validation/handler.go
@@ -1,29 +1,63 @@
+/*
+Copyright 2019 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package validation
 
 import (
 	"context"
-	"fmt"
+	"github.com/hashicorp/go-multierror"
 	sc "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/kubernetes-incubator/service-catalog/pkg/webhookutil"
-	"k8s.io/apimachinery/pkg/types"
+	admissionTypes "k8s.io/api/admission/v1beta1"
 	"net/http"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
+// Validator is used to implement new validation logic
+type Validator interface {
+	Validate(context.Context, admission.Request, *sc.ServiceBinding, *webhookutil.TracedLogger) error
+}
+
+// AdmissionHandler handles ServiceBinding validation
 type AdmissionHandler struct {
 	decoder *admission.Decoder
 	client  client.Client
+
+	CreateValidators []Validator
+	UpdateValidators []Validator
 }
 
 var _ admission.Handler = &AdmissionHandler{}
+var _ admission.DecoderInjector = &AdmissionHandler{}
+var _ inject.Client = &AdmissionHandler{}
 
-// Handle checks if instance reference for ServiceBinding is not marked for deletion
-// fail Create/Update ServiceBinding operation if the ServiceInstance is marked for deletion
+// NewAdmissionHandler creates new AdmissionHandler and initializes validators list
+func NewAdmissionHandler() *AdmissionHandler {
+	return &AdmissionHandler{
+		CreateValidators: []Validator{&ReferenceDeletion{}},
+		UpdateValidators: []Validator{&ReferenceDeletion{}},
+	}
+}
+
+// Handle handles admission requests.
 func (h *AdmissionHandler) Handle(ctx context.Context, req admission.Request) admission.Response {
 	traced := webhookutil.NewTracedLogger(req.UID)
-	traced.Infof("Start validation handling operation: %s for %s: %q", req.Operation, req.Kind.Kind, req.Name)
+	traced.Infof("Start handling AdmissionHandler operation: %s for %s: %q", req.Operation, req.Kind.Kind, req.Name)
 
 	sb := &sc.ServiceBinding{}
 	if err := webhookutil.MatchKinds(sb, req.Kind); err != nil {
@@ -36,40 +70,58 @@ func (h *AdmissionHandler) Handle(ctx context.Context, req admission.Request) ad
 		return admission.Errored(http.StatusBadRequest, err)
 	}
 
-	instanceRef := sb.Spec.InstanceRef
-	instance := &sc.ServiceInstance{}
+	var errs error
 
-	err := h.client.Get(ctx, types.NamespacedName{Namespace: sb.Namespace, Name: instanceRef.Name}, instance)
-	if err != nil {
-		traced.Errorf("Could not get ServiceInstance by name %q: %v", instanceRef.Name, err)
-		return admission.Errored(http.StatusBadRequest, err)
+	switch req.Operation {
+	case admissionTypes.Create:
+		for _, v := range h.CreateValidators {
+			if err := v.Validate(ctx, req, sb, traced); err != nil {
+				errs = multierror.Append(errs, err)
+			}
+		}
+	case admissionTypes.Update:
+		for _, v := range h.UpdateValidators {
+			if err := v.Validate(ctx, req, sb, traced); err != nil {
+				errs = multierror.Append(errs, err)
+			}
+		}
+	default:
+		traced.Infof("ServiceBinding AdmissionHandler wehbook does not support action %q", req.Operation)
+		return admission.Allowed("action not taken")
 	}
 
-	if instance.DeletionTimestamp != nil {
-		traced.Infof(
-			"Could not handle %s operation for %s because ServiceInstance %s is marked for deletion",
-			req.Operation,
-			req.Kind.Kind,
-			instanceRef.Name)
-		return admission.Denied(fmt.Sprintf("Could not %s %s %q", req.Operation, sb.Kind, sb.Name))
+	if errs != nil {
+		return admission.Denied(errs.Error())
 	}
 
-	traced.Infof("Completed successfully validation operation: %s for %s: %q", req.Operation, sb.Kind, sb.Name)
-	return admission.Allowed("Validation successful")
+	traced.Infof("Completed successfully AdmissionHandler operation: %s for %s: %q", req.Operation, req.Kind.Kind, req.Name)
+	return admission.Allowed("ServiceBinding AdmissionHandler successful")
 }
 
-var _ inject.Client = &AdmissionHandler{}
+// InjectDecoder injects the decoder into the handlers
+func (h *AdmissionHandler) InjectDecoder(d *admission.Decoder) error {
+	h.decoder = d
 
-// InjectClient injects the client into the AdmissionHandler
-func (h *AdmissionHandler) InjectClient(c client.Client) error {
-	h.client = c
+	for _, v := range h.CreateValidators {
+		admission.InjectDecoderInto(d, v)
+	}
+	for _, v := range h.UpdateValidators {
+		admission.InjectDecoderInto(d, v)
+	}
+
 	return nil
 }
 
-var _ admission.DecoderInjector = &AdmissionHandler{}
+// InjectClient injects the client into the handlers
+func (h *AdmissionHandler) InjectClient(c client.Client) error {
+	h.client = c
 
-// InjectDecoder injects the decoder into the AdmissionHandler
-func (h *AdmissionHandler) InjectDecoder(d *admission.Decoder) error {
-	h.decoder = d
+	for _, v := range h.CreateValidators {
+		inject.ClientInto(c, v)
+	}
+	for _, v := range h.UpdateValidators {
+		inject.ClientInto(c, v)
+	}
+
 	return nil
 }
